+++ pyomo/core/tests/unit/test_taylor_series.py
@@ -1,6 +1,6 @@
 import pyutilib.th as unittest
 import pyomo.environ as pe
-from pyomo.core.expr.taylor_series.taylor import taylor_series
+from pyomo.core.expr.taylor_series import get_taylor_series
 from pyomo.core.expr.current import polynomial_degree
 
 
@@ -11,7 +11,7 @@ class TestDerivs(unittest.TestCase):
         m.x.value = 1
         exprs_to_test = [m.x**2, pe.exp(m.x), (m.x + 2)**2]
         for e in exprs_to_test:
-            tsa = taylor_series(e)
+            tsa = get_taylor_series(e)
             self.assertAlmostEqual(pe.differentiate(e, wrt=m.x), pe.differentiate(tsa, wrt=m.x))
             self.assertAlmostEqual(pe.value(e), pe.value(tsa))
             self.assertEqual(polynomial_degree(tsa), 1)
