+++ pyomo/environ/__init__.py
@@ -106,5 +106,5 @@ from pyomo.opt import (
     TerminationCondition, SolverStatus,
     )
 from pyomo.core.base.units_container import units
-from pyomo.core.expr.differentiate.differentiate import differentiate, DiffModes
+from pyomo.core.expr.calculus.derivatives import differentiate, DiffModes
