+++ pyomo/util/tests/test_calc_var_value.py
@@ -15,7 +15,7 @@ import pyutilib.th as unittest
 from pyomo.common.log import LoggingIntercept
 from pyomo.environ import ConcreteModel, Var, Constraint, value, exp
 from pyomo.util.calc_var_value import calculate_variable_from_constraint
-from pyomo.core.expr.differentiate.diff_with_sympy import differentiate_available
+from pyomo.core.expr.calculus.diff_with_sympy import differentiate_available
 
 class Test_calc_var(unittest.TestCase):
     def test_initialize_value(self):
